Config-changed event, refactoring
* events module
* remove ComparableLocation, use Location<NoId> instead
* move LocationStats to separate module
