+++ src-tauri/src/events.rs
@@ -1,0 +1,6 @@
+pub static SINGLE_INSTANCE: &str = "single-instance";
+pub static CONNECTION_CHANGED: &str = "connection-changed";
+pub static INSTANCE_UPDATE: &str = "instance-update";
+pub static LOCATION_UPDATE: &str = "location-update";
+pub static APP_VERSION_FETCH: &str = "app-version-fetch";
+pub static CONFIG_CHANGED: &str = "config-changed";
