+++ src-tauri/src/lib.rs
@@ -4,6 +4,7 @@ pub mod appstate;
 pub mod commands;
 pub mod database;
 pub mod error;
+pub mod events;
 pub mod periodic;
 pub mod service;
 pub mod tray;
