+++ src-tauri/src/periodic/config.rs
@@ -5,46 +5,61 @@ use tracing::warn;
 
 use crate::{
     appstate::AppState,
-    commands::{device_config_to_location, save_device_config},
-    database::{models::Id, DbPool, Instance, Location, WireguardKeys},
+    commands::{device_config_to_location, update_instance},
+    database::{
+        models::{Id, NoId},
+        DbPool, Instance, Location, WireguardKeys,
+    },
     error::Error,
-    proto::{DeviceConfig, InstanceInfoRequest, InstanceInfoResponse},
+    events::{CONFIG_CHANGED, INSTANCE_UPDATE},
+    proto::{InstanceInfoRequest, InstanceInfoResponse},
 };
 
 const INTERVAL_SECONDS: Duration = Duration::from_secs(30);
 const POLLING_ENDPOINT: &str = "api/v1/poll";
 
-// pub async fn check_config(app_handle: AppHandle, instances: Vec<Instance>) {
-pub async fn check_config(handle: AppHandle) {
+/// Periodically retrieves and updates configuration for all [`Instance`]s.
+/// Updates are only performed if no connections are established for the [`Instance`].
+pub async fn poll_config(handle: AppHandle) {
     let state: State<AppState> = handle.state();
     let pool = state.get_pool();
     loop {
+        // TODO(jck): unwrap
         let instances = Instance::all(&pool).await.unwrap();
-        debug!("Polling for configuration updates",);
+        debug!(
+            "Polling configuration updates for {} instances",
+            instances.len(),
+        );
         for instance in &instances {
-            match update_instance_config(&pool, instance, &state, &handle).await {
-                Ok(_config) => println!("TEST TEST"),
-                Err(err) => error!("Failed to fetch instance {} config, {}", instance.name, err),
+            if let Err(err) = poll_instance(&pool, instance, handle.clone()).await {
+                error!(
+                    "Failed to retrieve instance {}({}) config, {}",
+                    instance.name, instance.id, err
+                );
+            } else {
+                debug!(
+                    "Retrieved config for instance {}({})",
+                    instance.name, instance.id,
+                );
             }
         }
-
         info!(
-            "Retrieved configuration, sleeping {}s",
-            INTERVAL_SECONDS.as_secs()
+            "Retrieved configuration for {} instances, sleeping {}s",
+            instances.len(),
+            INTERVAL_SECONDS.as_secs(),
         );
         sleep(INTERVAL_SECONDS).await;
-
-        // let response = get_latest_app_version(app_handle.clone()).await;
     }
 }
 
-async fn update_instance_config(
+/// Retrieves configuration for given [`Instance`].
+/// Updates the instance if there are no active connections, otherwise displays UI message.
+async fn poll_instance(
     pool: &DbPool,
     instance: &Instance<Id>,
-    state: &State<'_, AppState>,
-    _handle: &AppHandle,
+    handle: AppHandle,
 ) -> Result<(), Error> {
-    // TODO(jck): unwraps
+    // TODO(jck): unwrap
     let WireguardKeys { pubkey, .. } = WireguardKeys::find_by_instance_id(pool, instance.id)
         .await?
         .unwrap();
@@ -61,12 +76,10 @@ async fn update_instance_config(
         token: token.clone(),
         pubkey,
     };
-    // let response = client.get(url).header("defguard_proxy", TOKEN).send().await;
     let response = client.post(url).json(&request).send().await;
     if let Ok(response) = response {
         info!("Reponse: {:#?}", response);
         // TODO(jck): unwrap
-        // CONTINUE: no uuid present...
         let response: InstanceInfoResponse = response.json().await.unwrap();
         // TODO(jck): unwraps
         let instance = Instance::find_by_uuid(
@@ -83,26 +96,25 @@ async fn update_instance_config(
         // TODO(jck): unwrap
         .await?
         .unwrap();
-        // TODO(jck): unwrap
         let db_locations = Location::find_by_instance_id(pool, instance.id).await?;
-        let db_comparable_locations: Vec<ComparableLocation> = db_locations
+        // let db_comparable_locations: Vec<ComparableLocation> = db_locations
+        let db_comparable_locations: Vec<Location<NoId>> = db_locations
             .into_iter()
-            .map(ComparableLocation::from)
+            .map(Location::<NoId>::from)
             .collect();
-        let db_comparable_locations: HashSet<ComparableLocation> =
+        let db_comparable_locations: HashSet<Location<NoId>> =
             HashSet::from_iter(db_comparable_locations);
         // TODO(jck): unwrap
-        let core_locations: Vec<ComparableLocation> = response
+        let core_locations: Vec<Location<NoId>> = response
             .device_config
             .as_ref()
             .unwrap()
             .configs
             .iter()
             .map(|config| device_config_to_location(config.clone(), instance.id))
-            .map(ComparableLocation::from)
+            .map(Location::<NoId>::from)
             .collect();
-        let core_comparable_locations: HashSet<ComparableLocation> =
-            HashSet::from_iter(core_locations);
+        let core_comparable_locations: HashSet<Location<NoId>> = HashSet::from_iter(core_locations);
 
         if core_comparable_locations == db_comparable_locations {
             // config remains unchanged, return early
@@ -111,59 +123,33 @@ async fn update_instance_config(
 
         // Config changed
 
-        // If there are no active connections for this instance, then update the database.
+        // If there are no active connections for this instance, update the database.
         // Otherwise just display the message to reconnect.
+        let state: State<'_, AppState> = handle.state();
         if state.active_connections(&instance).await?.is_empty() {
-            todo!();
+            // TODO(jck): unwrap
+            debug!(
+                "Updating instance {}({}) configuration: {:?}",
+                instance.name, instance.id, response.device_config,
+            );
+            update_instance(
+                instance.id,
+                response.device_config.unwrap(),
+                state,
+                handle.clone(),
+            )
+            .await?;
+            handle.emit_all(INSTANCE_UPDATE, ())?;
+            info!(
+                "Updated instance {}({}) configuration",
+                instance.name, instance.id
+            );
         } else {
-            todo!();
+            let _ = handle.emit_all(CONFIG_CHANGED, &instance.name);
         }
 
-        // // TODO(jck): unwrap
-        // let _ = save_device_config(
-        //     prvkey,
-        //     token,
-        //     response.device_config.unwrap(),
-        //     state.clone(),
-        //     handle.clone(),
-        // )
-        // .await;
         Ok(())
     } else {
         Err(Error::InternalError("TODO(jck)".to_string()))
-    }
-}
-
-#[derive(Hash, Debug, Clone, PartialEq, Eq)]
-struct ComparableLocation {
-    pub instance_id: i64,
-    // Native id of network from defguard
-    pub network_id: i64,
-    pub name: String,
-    pub address: String,
-    pub pubkey: String,
-    pub endpoint: String,
-    pub allowed_ips: String,
-    pub dns: Option<String>,
-    pub route_all_traffic: bool,
-    pub mfa_enabled: bool,
-    pub keepalive_interval: i64,
-}
-
-impl<I> From<Location<I>> for ComparableLocation {
-    fn from(location: Location<I>) -> Self {
-        Self {
-            instance_id: location.instance_id,
-            network_id: location.network_id,
-            name: location.name,
-            address: location.address,
-            pubkey: location.pubkey,
-            endpoint: location.endpoint,
-            allowed_ips: location.allowed_ips,
-            dns: location.dns,
-            route_all_traffic: location.route_all_traffic,
-            mfa_enabled: location.mfa_enabled,
-            keepalive_interval: location.keepalive_interval,
-        }
     }
 }
